# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM1 API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/change-password": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Change password",
                "operationId": "76ac7e45dd3e106d2284cca1169eee99",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ResetPasswordRequest"
                },
                "responses": {
                    "200": {
                        "description": "Change success"
                    },
                    "500": {
                        "description": "Error server"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/disable-user": {
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Лишение прав",
                "operationId": "9067e16cea659784b39ef0f64b3d12fe",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateUserRequest"
                },
                "responses": {
                    "200": {
                        "description": "Desable success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/change-role": {
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Смена роли пользователя",
                "operationId": "984a75a7fa8578c8814c686ee9f68bff",
                "requestBody": {
                    "$ref": "#/components/requestBodies/ChangeRoleRequest"
                },
                "responses": {
                    "200": {
                        "description": "Change role success"
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/index": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Список пользователей",
                "operationId": "b275490f7e197988d85ca38eeb7cac8e",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error server"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Поиск по id",
                "operationId": "e5872d1157a1c324c1806ad3a2fb8829",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Удаление по id",
                "operationId": "976d880c54adeccb085a8f1bd1dc78cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/by-email": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Поиск по id",
                "operationId": "9ba7d9e9fd7ba23eb2c36db0d5f09d15",
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateUserRequest"
                },
                "responses": {
                    "200": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Создание пользователя",
                "operationId": "faa1e9e5665314af7c7ccd0dcf907ba7",
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateUserRequest"
                },
                "responses": {
                    "201": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/update/{id}": {
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "Обновление по id",
                "operationId": "20ac078d83824c5db2d30e42ddbcada4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateUserRequest"
                },
                "responses": {
                    "200": {
                        "description": "User resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "name",
                    "email",
                    "is_active"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Adam"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "t@test.com"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "ChangeRoleRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "email",
                                "role"
                            ],
                            "properties": {
                                "email": {
                                    "description": "email для поиска user",
                                    "type": "string",
                                    "format": "email",
                                    "example": "user@example.com"
                                },
                                "role": {
                                    "description": "Доступрые роли: admin, manager, user, viewer",
                                    "type": "string",
                                    "example": "manager"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "CreateUserRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "name",
                                "email",
                                "password",
                                "is_active",
                                "role"
                            ],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Adam"
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email",
                                    "example": "t@test.com"
                                },
                                "password": {
                                    "description": "Minimum 8 symbols",
                                    "type": "string",
                                    "example": "test00000"
                                },
                                "is_active": {
                                    "type": "boolean",
                                    "example": "true"
                                },
                                "role": {
                                    "description": "Из списка: admin, manager, user, viewer",
                                    "type": "string",
                                    "example": "user"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "ResetPasswordRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "current_password",
                                "new_password"
                            ],
                            "properties": {
                                "current_password": {
                                    "description": "Minimum 8 symbols",
                                    "type": "string",
                                    "example": "test11111"
                                },
                                "new_password": {
                                    "description": "Minimum 8 symbols",
                                    "type": "string",
                                    "example": "test00000"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateUserRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [],
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Adam",
                                    "nullable": true
                                },
                                "email": {
                                    "type": "string",
                                    "format": "email",
                                    "example": "t@test.com",
                                    "nullable": true
                                },
                                "password": {
                                    "description": "Minimum 8 symbols",
                                    "type": "string",
                                    "example": "test00000",
                                    "nullable": true
                                },
                                "is_active": {
                                    "type": "boolean",
                                    "example": "true",
                                    "nullable": true
                                },
                                "role": {
                                    "description": "Из списка: admin, manager, user, viewer",
                                    "type": "string",
                                    "example": "user",
                                    "nullable": true
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}